+++ lib/src/utils/data_list_controller.dart
@@ -1,5 +1,6 @@
 import '../screen_controller.dart';
-import './data_controller.dart';
+
+import 'data_controller.dart';
 
 /// This is same as [DataScreenController], except that it is specialized for list
 /// T is the type of list item, so if T is String, then list will be type of List<String>